Move scheduler back into streaming_vocoder.py and apply review cleanups
Reverts the previous commit's relocation of HiggsStreamingVocoderScheduler
into stages.py — the scheduler stays in its own module, which keeps the
factory file small and matches the s2-pro layout.

On the streaming_vocoder side: drop the write-only audio_chunks state
in favor of a has_emitted bool (the only consumer was a truthiness
check), drop the .cpu() in _decode_delayed_rows so audio stays on the
codec device until the single serialization call, and drop the
secondary "stream"-metadata gate now that the engine is the single
source of truth for whether to emit chunks.

On the stages side: drop the hasattr guard around set_stream_outbox —
the HiggsTTSModelRunner is constructed two lines above and definitely
has the method.
